signal/main: Merge signal into main.
