+++ CMakeLists.txt
@@ -9,7 +9,6 @@ pkg_check_modules(NCURSES REQUIRED ncurses)
 
 set(WW_HEADERS
     "common.h"
-    "signal.h"
     "console_win.h"
     "reactor_mgr.h"
     "reactor.h"
@@ -21,7 +20,6 @@ set(WW_SOURCES
     "common.c"
     "console_win.c"
     "main.c"
-    "signal.c"
     "reactor_mgr.c"
     "reactor.c"
     "status_win.c"
