--- signal.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void signal_setup_handlers(void);
